+++ src/cmd/shuffle.rs
@@ -6,25 +6,19 @@ use crate::util;
 use crate::CliResult;
 
 static USAGE: &str = "
-Shuffle the given CSV file. Requires memory proportional to the
-number of rows of the file (approx. 2 u64 per row).
-
-Note that rows from input file are copied as-is in the output.
-This means that no CSV serialization harmonization will happen,
-unless --in-memory is set.
-
-Also, since this command needs random access in the input file, it
-does not work with stdin or piping (unless --in-memory) is set.
+Shuffle the rows of a given CSV file. This requires loading the whole
+file in memory. If memory is scarce and target file is seekable (not stdin,
+nor unindexed compressed file), you can also use the -e/--external flag that
+only requires memory proportional to the number of rows of the file.
 
 Usage:
     xan shuffle [options] [<input>]
     xan shuffle --help
 
 shuffle options:
-    --seed <number>        RNG seed.
-    -m, --in-memory        Load all CSV data in memory before shuffling it. Can
-                           be useful for streamed inputs such as stdin but of
-                           course costs more memory.
+    --seed <number>  RNG seed.
+    -e, --external   Shuffle the file without buffering it into memory. Only
+                     works if target is seekable (no stdin etc.).
 
 Common options:
     -h, --help             Display this message
@@ -42,16 +36,16 @@ struct Args {
     flag_no_headers: bool,
     flag_delimiter: Option<Delimiter>,
     flag_seed: Option<usize>,
-    flag_in_memory: bool,
+    flag_external: bool,
 }
 
-fn run_random_access(args: Args) -> CliResult<()> {
+fn run_external(args: Args) -> CliResult<()> {
     let rconf = Config::new(&args.arg_input)
         .delimiter(args.flag_delimiter)
         .no_headers(args.flag_no_headers);
     let wconf = Config::new(&args.flag_output);
 
-    // Seeding rng
+    // Seedmemorying rng
     let mut rng = util::acquire_rng(args.flag_seed);
 
     let mut header_len: Option<usize> = None;
@@ -141,9 +135,9 @@ fn run_in_memory(args: Args) -> CliResult<()> {
 pub fn run(argv: &[&str]) -> CliResult<()> {
     let args: Args = util::get_args(USAGE, argv)?;
 
-    if args.flag_in_memory {
-        run_in_memory(args)
+    if args.flag_external {
+        run_external(args)
     } else {
-        run_random_access(args)
+        run_in_memory(args)
     }
 }